refactor(server): find flat serverless function util
